refactor(cli): colocate index/stats schema errors and exit helpers
Move SCHEMA_INPUT_ERRORS, exit_schema_input_error, and
exit_redis_search_error from utils into index.py and stats.py so CLI
behavior is visible next to each command. utils keeps URL and argparse
helpers only.

Made-with: Cursor
